+++ src/pages/Cart/Cart.scss
@@ -9,8 +9,7 @@
     // @include flexCenter;
     font-weight: $font-weight-semibold;
   }
-
-  .cartItemsCounts {
+  .cartItemsCount {
     margin: 25px 0;
     font-size: $font-size-sm;
     color: $font-color-grey;
@@ -29,131 +28,20 @@
       padding: 15px 0px;
       border-top: 1px solid $background-color-darkgrey;
       border-bottom: 1px solid $background-color-darkgrey;
-    }
+    } //리스트 머리
+
     .cartItemsList {
       display: flex;
       flex-direction: column;
-
-      .cartItem {
-        @include flexSort(null, flex-start);
-        flex: 1 1 auto;
-        width: 100%;
-        padding: 15px 25px 30px 25px;
-        border-bottom: 1px solid $background-color-darkgrey;
-
-        .cartItemImg {
-          width: 130px;
-          height: 130px;
-          margin-right: 15px;
-        }
-        .cartItemInfo {
-          font-size: $font-size-sm;
-
-          .cartItemTitle {
-            margin-bottom: 10px;
-            font-weight: $font-weight-semibold;
-          }
-          .cartItemDes {
-            line-height: 1.2em;
-          }
-        }
-        .cartOptChange {
-          flex: 1 1 auto;
-        }
-        .cartItemPrice {
-          margin-right: 20px;
-          color: $font-color-red;
-          font-weight: $font-weight-semibold;
-        }
-      }
-
-      .addWishBtn {
-        color: $font-color-grey;
-        margin-left: 25px;
-      }
     }
   }
 }
-.cartRight {
-  width: 350px;
 
-  .cartRightWrapper {
-    flex-direction: column;
-    border: 1px, solid $background-color-darkgrey;
-
-    .checkoutTitle {
-      background-color: $background-color-darkgrey;
-      padding: 17px 0 17px 10px;
-      font-weight: $font-weight-bold;
-      font-size: $font-size-sm;
-    }
-
-    .checkoutDetail {
-      font-size: $font-size-sm;
-      line-height: 2.5em;
-      padding: 17px;
-      border: 1px solid $background-color-darkgrey;
-      font-weight: $font-weight-semibold;
-
-      .detail {
-        display: flex;
-        justify-content: space-between;
-        font-size: $font-size-xs;
-      }
-
-      .totalPrice {
-        margin-top: 0.5em;
-        font-size: $font-size-lg;
-        font-weight: $font-weight-semibold;
-      }
-
-      .checkoutOrderBtn {
-        width: 100%;
-        margin-top: 0.8em;
-        padding: 10px 0;
-        background-color: $font-color-red;
-        color: rgb(255, 255, 255);
-      }
-
-      //오른쪽박스}
-      .coupon {
-        margin-top: 2em;
-        font-weight: $font-weight-semibold;
-        font-size: $font-size-xs;
-
-        .couponNum {
-          text-decoration: underline;
-          color: $font-color-red;
-        }
-      }
-
-      .couponDes {
-        margin: 1em 0 0 1em;
-        color: $font-color-grey;
-        font-size: 8px;
-      }
-
-      .inputCoupon {
-        height: 50px;
-        width: 100%;
-        margin-top: 25px;
-        border: 1px solid $background-color-darkgrey;
-
-        .couponCode {
-          border: none;
-          outline: none;
-          width: 70%;
-          font-size: $font-size-xs;
-        }
-
-        .couponBtn {
-          width: 30%;
-          height: 50px;
-          border-left: 1px solid $background-color-darkgrey;
-          justify-content: center;
-          text-align: center;
-        }
-      }
-    } //쿠폰
-  } //오른쪽
-} //장바구니 감싸는거
+.optOverlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  height: 100vh;
+  background-color: rgba(0, 0, 0, 0.8);
+}
