+++ src/pages/Cart/Commponent/CartAside.scss
@@ -1,0 +1,74 @@
+@import '../../../styles/variables.scss';
+@import 'https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.1.2/css/all.min.css';
+
+.cartAside {
+  width: 350px;
+
+  .cartCheckoutWrapper {
+    flex-direction: column;
+    border: 1px, solid $background-color-darkgrey;
+
+    .checkoutTitle {
+      background-color: $background-color-darkgrey;
+      padding: 17px 0 17px 10px;
+      font-weight: $font-weight-bold;
+      font-size: $font-size-sm;
+    }
+
+    .checkoutDetail {
+      font-size: $font-size-sm;
+      line-height: 2.5em;
+      padding: 17px;
+      border: 1px solid $background-color-darkgrey;
+      font-weight: $font-weight-semibold;
+
+      .detail {
+        display: flex;
+        justify-content: space-between;
+        font-size: $font-size-xs;
+      }
+    }
+
+    .totalPrice {
+      margin-top: 0.5em;
+      font-size: $font-size-lg;
+      font-weight: $font-weight-semibold;
+    }
+
+    .checkoutOrderBtn {
+      width: 100%;
+      margin-top: 0.8em;
+      padding: 10px 0;
+      background-color: $font-color-red;
+      color: rgb(255, 255, 255);
+    }
+  }
+
+  .coupon {
+    margin-top: 2em;
+    font-weight: $font-weight-semibold;
+    font-size: $font-size-xs;
+
+    .inputCoupon {
+      height: 50px;
+      width: 100%;
+      margin-top: 25px;
+      border: 1px solid $background-color-darkgrey;
+
+      .couponCode {
+        border: none;
+        outline: none;
+        width: 70%;
+        font-size: $font-size-xs;
+      }
+
+      .couponBtn {
+        width: 30%;
+        height: 50px;
+        border-left: 1px solid $background-color-darkgrey;
+        justify-content: center;
+        text-align: center;
+      }
+    }
+  }
+}
