+++ src/pages/Cart/Commponent/CartItem.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables.scss';
+@import '../../../styles/variables.scss';
 @import 'https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.1.2/css/all.min.css';
 
 .cartItem {
@@ -12,7 +12,9 @@
     width: 130px;
     height: 130px;
     margin-right: 15px;
+    cursor: pointer;
   }
+
   .cartItemInfo {
     font-size: $font-size-sm;
 
@@ -20,16 +22,44 @@
       margin-bottom: 10px;
       font-weight: $font-weight-semibold;
     }
+
     .cartItemDes {
       line-height: 1.2em;
     }
   }
+
   .cartOptChange {
     flex: 1 1 auto;
   }
+
   .cartItemPrice {
-    margin-right: 20px;
-    color: $font-color-red;
-    font-weight: $font-weight-semibold;
+    line-height: 1.4em;
+
+    .cartRetailPrice {
+      margin-right: 20px;
+      color: $font-color-red;
+      font-weight: $font-weight-semibold;
+    }
+
+    .beforePrice {
+      display: flex;
+      position: relative;
+
+      &::after {
+        content: '';
+        position: absolute;
+        top: 50%;
+        border-bottom: 1px solid $background-color-black;
+        width: 3.6em;
+      }
+    }
+    .cartPriceNone {
+      display: none;
+    }
+  }
+
+  .addWishBtn {
+    color: $font-color-grey;
+    margin-left: 25px;
   }
 }
