--- src/pages/Commponent/CartNull.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function CartNull() {
-  return <div>장바구니</div>;
-}
-
-export default CartNull;
